feat(swingset): compose password dialog in profile stories
